ai-engine/requirements.txt
requests>=2.28
pydantic>=1.10
pymongo>=4.6
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
faiss-cpu; platform_system != 'Windows' or platform_system == 'Windows'
numpy>=1.24
beautifulsoup4>=4.12
python-pptx>=1.0
python-docx>=1.0
spacy>=3.5
joblib>=1.3
PyPDF2>=3.0.0
openai>=1.0.0
sentence-transformers>=2.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
